Rearchitect getPossibleSteals to optimize runtime
This commit improves getPossibleSteals('test') from ~18 to ~0.8 seconds.

I rewrote the code to frontload computing, so that getDict does the
computation that was previously recomputed on every call to
getPossibleSteals.

Previously, getDict() returned a set containing every word from
dictionaryFile. Now, it returns a dict mapping a length to a list of
words of that length. Each list contains words as LibWord() objects,
which contain (precomputed) the root of the word and a Counter object
for the letters.

getPossibleSteals now also only looks at words of the appropriate
length, cutting out words shorter than the input or longer than the max.